seperated the HOC components into different folder
